pyproject.toml
[project]
name = "secrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "anthropic>=0.84.0",
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.135.1",
    "greenlet>=3.3.2",
    "langchain>=1.2.10",
    "psycopg2-binary>=2.9.11",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.0",
    "sqlalchemy>=2.0.48",
    "strawberry-graphql>=0.311.1",
    "uvicorn>=0.41.0",
    "voyageai>=0.3.7",
]

[tool.uv]
package = true

[tool.setuptools.packages.find]
where = ["src"]
